This is mostly an interface cleanup to provide some improvements to look and feel & (hopefully) useability.  This is really just the "eye candy" release... with the usual caveats about bugs.
The visualisation component is bigger (800x800) and now uses a white background (hopefully a cleaner feel).  Most notably, there is a very early prototype of a node menu now included in the visualisation component, which should give you an idea of where things are heading.  I've got the delete function working via RESTful routes, but nothing else.  As you will see, there is stacks of work to be done on event handlers to make this component fluid and easy to use, but I thought it was worth pushing out for people to see what I've been working on...

I've added a couple of Javascript libraries too - JQueryUI and a little one called NiftyCube (not displaying too well in IE). This means nicer widgets like the date picker and the accordion for grouped lists.

Signed-off-by: sashaagafonoff <sasha.agafonoff@gmail.com>